chore(structure): move app-header inside its own folder
